api/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
google-cloud-bigquery>=3.17.0
google-cloud-aiplatform>=1.40.0
pydantic>=2.6.0
python-dotenv>=1.0.0
